Allow scripts to work if laravel project exists in subdirectory (#45)
* Run commands from current directory

Instead of forcing commands to be from repository
root, let's work from the current directory. This
allows the laravel instance to reside in another
directory within the project rather than always
be in the root directory.

* Set root dir for nginx to public within current dir

Similiarily to commit d5871b6, this allows for more
flexibility in where the laravel resides within
the project.

* Make PHP-FPM run as default CMD

Since you can not have two CMDs and we want to
reconfigure nginx before starting it, let's make
the image rely on PHP-FPM.

* Fix CMD command

php7.2-fpm does not exist and its parameter
is not correct for what it seems to be for.